sentiment-analyzer/requirements.txt
kafka-python
nltk
beautifulsoup4
psycopg2-binary